requirements.txt
Django==1.10.6
django-ckeditor==5.2.1
django-flat-responsive==1.3.0
django-widget-tweaks==1.4.1
geoip2==2.4.2
gunicorn==19.7.0
maxminddb==1.2.3
olefile==0.44
Pillow==4.0.0
pkg-resources==0.0.0
requests==2.13.0
